% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliaries.R
\name{draw_dens_poseidon}
\alias{draw_dens_poseidon}
\title{Title}
\usage{
draw_dens_poseidon(
  VV,
  k = 1,
  xlim = NULL,
  n_points_dens = 5000,
  hist_breaks = 25
)
}
\arguments{
\item{brbr}{}
}
\description{
Title
}
