% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistart.R
\name{extract_best_poseidon}
\alias{extract_best_poseidon}
\title{Extract best run from a list of Poseidons}
\usage{
extract_best_poseidon(VVm, min = NULL, plot = FALSE)
}
\arguments{
\item{min}{}
}
\description{
Extract best run from a list of Poseidons
}
