% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_biclust.R
\name{plot_bicluster_poseidon}
\alias{plot_bicluster_poseidon}
\title{Title}
\usage{
plot_bicluster_poseidon(poseidon_fit, data = 1)
}
\arguments{
\item{data}{...}

\item{q}{...}
}
\value{
...
}
\description{
Title
}
