% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POS_fiSAN.R
\name{POSEIDON_fiSAN}
\alias{POSEIDON_fiSAN}
\title{Title}
\usage{
POSEIDON_fiSAN(
  Y_list,
  L,
  K,
  NN_inds,
  XIi,
  RHOi,
  itemp = 0,
  itemp_grid = seq(0, 2, by = 0.1),
  nsim = 50,
  replicates_potts = 5,
  epsilon = 1e-04,
  hyperpar,
  seed = NULL,
  verbose = TRUE,
  estimate_itemp = FALSE,
  compute_elbo_every = 1
)
}
\arguments{
\item{Y_list}{...}

\item{L}{...}

\item{K}{...}

\item{NN_inds}{...}

\item{itemp}{...}

\item{itemp_grid}{...}

\item{nsim}{...}

\item{replicates_potts}{...}

\item{epsilon}{...}

\item{hyperpar}{...}

\item{seed}{...}

\item{verbose}{...}

\item{estimate_itemp}{...}

\item{compute_elbo_every}{...}

\item{warmstart}{...}

\item{IN}{...}
}
\value{
...
}
\description{
Title
}
\examples{
1+1
}
