% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POS_onlyfiSAN.R
\name{POSEIDON_onlyfiSAN}
\alias{POSEIDON_onlyfiSAN}
\title{Title}
\usage{
POSEIDON_onlyfiSAN(
  Y_list,
  L,
  K,
  XIi,
  RHOi,
  nsim = 50,
  epsilon = 1e-04,
  hyperpar,
  seed = NULL,
  verbose = TRUE,
  compute_elbo_every = 1
)
}
\arguments{
\item{Y_list}{...}

\item{L}{...}

\item{K}{...}

\item{nsim}{...}

\item{epsilon}{...}

\item{hyperpar}{...}

\item{seed}{...}

\item{verbose}{...}

\item{compute_elbo_every}{...}

\item{NN_inds}{...}

\item{replicates_potts}{...}

\item{warmstart}{...}

\item{itemp}{...}

\item{itemp_grid}{...}

\item{IN}{...}

\item{estimate_itemp}{...}
}
\value{
...
}
\description{
Title
}
\examples{
1+1
}
