% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{random_init_poseidon}
\alias{random_init_poseidon}
\title{Title}
\usage{
random_init_poseidon(Y, L, K, seed = NULL, hyperpar, warmstart, BY = 0.1)
}
\arguments{
\item{Y}{...}

\item{L}{...}

\item{K}{...}

\item{seed}{...}

\item{hyperpar}{...}

\item{warmstart}{...}

\item{BY}{...}
}
\value{
...
}
\description{
Title
}
\examples{
 1+1
}
